fix(core): fix CGFloat type conversions in ImageBlender and ImageStitchingEngine

+++ LongScreenshot/Extensions/CGImage+Crop.swift
@@ -4,7 +4,7 @@ import UIKit
 public extension CGImage {
     /// Safely crops a CGImage to a sub-rectangle within bounds
     func safeCropping(to rect: CGRect) -> CGImage? {
-        let imageBounds = CGRect(x: 0, y: 0, width: self.width, height: self.height)
+        let imageBounds = CGRect(x: 0, y: 0, width: CGFloat(self.width), height: CGFloat(self.height))
         let intersection = imageBounds.intersection(rect)
         
         guard !intersection.isNull, intersection.width > 0, intersection.height > 0 else {
